pyproject.toml
[project]
name = "stock-selection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "finmind>=1.9.7",
    "google-generativeai>=0.8.6",
    "loguru>=0.7.3",
    "lxml>=6.1.0",
    "pandas>=3.0.2",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "tqdm>=4.67.3",
    "yfinance>=1.3.0",
]
